+++ ChangeLog.md
@@ -5,3 +5,4 @@
 - Optimized codes
 - Added screen break points checker
 - Added universal style getter 
+- Added browser Resize Property handler